Add color of target rank to /util response embed
